fix pipeline issue

+++ Jenkinsfile
@@ -5,6 +5,7 @@ pipeline {
         MINOR_BUILD = 0
         BACK_IMAGE_NAME = "${env.GIT_BRANCH.toLowerCase()}-backend:${env.MAJOR_BUILD}.${env.MINOR_BUILD}.${env.BUILD_ID}"
         FRONT_IMAGE_NAME = "${env.GIT_BRANCH.toLowerCase()}-frontend:${env.MAJOR_BUILD}.${env.MINOR_BUILD}.${env.BUILD_ID}"
+        COMPOSE_STATUS = false
     }
     stages {
         stage('Prepering environment') {
@@ -39,9 +40,10 @@ pipeline {
         stage('Raise dockers environment'){
             steps{
                 withEnv(["BACK_IMAGE_NAME=${env.BACK_IMAGE_NAME}", "FRONT_IMAGE_NAME=${env.FRONT_IMAGE_NAME}"]) {
-                    sh 'printenv | grep IMAGE_NAME'
                     sh 'docker-compose up -d'
                 }
+                env.COMPOSE_STATUS = true
+                
             }
         }
         stage('Post environment stage checks'){
@@ -56,10 +58,12 @@ pipeline {
         }
     }
     post {
-        success{
-            sh "docker-compose down"
-        }
         cleanup {
+            script {
+                if (env.COMPOSE_STATUS) {
+                    sh "docker-compose down"
+                }
+            }
             sh "docker rmi -f ${env.BACK_IMAGE_NAME}"
             sh "docker rmi -f ${env.FRONT_IMAGE_NAME}"
             cleanWs()
